+++ examples/djangoapp/tests/app.py
@@ -43,7 +43,8 @@ class TestDjangoChat(unittest.TestCase):
                 '--bind', '127.0.0.1:0',
                 '--concurrency', cls.concurrency,
                 '--exc-id', cls.exc_id,
-                '--pulse-app-name', name]
+                '--pulse-app-name', name,
+                '--data-store', cls.data_store()]
         cls.app = yield send('arbiter', 'run', start_server, name, argv)
         assert cls.app.cfg.exc_id == cls.exc_id, "Bad execution id"
         cls.uri = 'http://{0}:{1}'.format(*cls.app.address)
@@ -55,6 +56,10 @@ class TestDjangoChat(unittest.TestCase):
         if cls.app:
             return send('arbiter', 'kill_actor', cls.app.name)
 
+    @classmethod
+    def data_store(cls):
+        return 'pulsar://127.0.0.1:0/8'
+
     def test_home(self):
         result = yield self.http.get(self.uri)
         self.assertEqual(result.status_code, 200)
@@ -63,18 +68,19 @@ class TestDjangoChat(unittest.TestCase):
         result = yield self.http.get('%s/bsjdhcbjsdh' % self.uri)
         self.assertEqual(result.status_code, 404)
 
-    def test_handshake(self):
-        ws = yield self.http.get(self.ws)
+    def test_websocket(self):
+        ws = yield self.http.get(self.ws, websocket_handler=MessageHandler())
         response = ws.handshake
         self.assertEqual(response.status_code, 101)
         self.assertEqual(response.headers['upgrade'], 'websocket')
         self.assertEqual(response.connection, ws.connection)
         self.assertTrue(ws.connection)
-
-    def test_websocket(self):
-        ws = yield self.http.get(self.ws, websocket_handler=MessageHandler())
-        self.assertTrue(ws)
         self.assertIsInstance(ws.handler, MessageHandler)
+        #
+        data = yield ws.handler.get()
+        data = json.loads(data)
+        self.assertEqual(data['message'], 'joined')
+        #
         ws.write('Hello there!')
         data = yield ws.handler.get()
         data = json.loads(data)